refactor(gax): consolidate grpc-rust test mocks
